+++ wazo_dird/database/queries/profile.py
@@ -4,6 +4,7 @@
 from sqlalchemy import (
     and_,
     exc,
+    func,
     text,
 )
 
@@ -26,52 +27,94 @@ class ProfileCRUD(BaseDAO):
 
     _profile_schema = schemas.ProfileSchema()
 
-    def create(self, body):
+    def count(self, visible_tenants, **list_params):
+        filter_ = self._list_filter(visible_tenants, **list_params)
         with self.new_session() as s:
-            self._create_tenant(s, body['tenant_uuid'])
+            query = s.query(func.count(Profile.uuid)).filter(filter_)
+            return query.scalar()
+
+    def create(self, body):
+        tenant_uuid = body['tenant_uuid']
+        with self.new_session() as s:
+            self._create_tenant(s, tenant_uuid)
             body['uuid'] = self._create_profile(s, body)
-            for service_name, config in body['services'].items():
-                service_uuid = self._create_service(s, service_name)
-                profile_service_uuid = self._create_profile_service(
-                    s, body['uuid'], service_uuid, config,
-                )
-                for source in config['sources']:
-                    source_uuid = source.get('uuid')
-                    self._create_profile_service_source(
-                        s, profile_service_uuid, source_uuid,
-                    )
+            self._associate_all_services(s, tenant_uuid, body['uuid'], body['services'])
         return body
 
     def delete(self, visible_tenants, profile_uuid):
-        filter_ = Profile.uuid == profile_uuid
-        if visible_tenants is not None:
-            if not visible_tenants:
-                raise exception.NoSuchProfile(profile_uuid)
-            filter_ = and_(filter_, Profile.tenant_uuid.in_(visible_tenants))
-
+        filter_ = self._build_filter(visible_tenants, profile_uuid)
         with self.new_session() as s:
             nb_deleted = s.query(Profile).filter(filter_).delete(synchronize_session=False)
 
         if not nb_deleted:
-            raise exception.NoSuchProfile(profile_uuid)
+            raise exception.NoSuchProfileAPIException(profile_uuid)
+
+    def edit(self, visible_tenants, profile_uuid, body):
+        filter_ = self._build_filter(visible_tenants, profile_uuid)
+        with self.new_session() as s:
+            self._dissociate_all_services(s, profile_uuid)
+
+            profile = s.query(Profile).filter(filter_).first()
+            if not profile:
+                raise exception.NoSuchProfileAPIException(profile_uuid)
+
+            profile.name = body['name']
+            profile.display_uuid = body['display']['uuid']
+            profile.display_tenant_uuid = profile.tenant_uuid
+            try:
+                s.flush()
+            except exc.IntegrityError as e:
+                constraint_name = extract_constraint_name(e)
+                if constraint_name == 'dird_profile_display_uuid_tenant_fkey':
+                    raise exception.NoSuchDisplay(body['display']['uuid'])
+                elif constraint_name == 'dird_profile_tenant_name':
+                    raise exception.DuplicatedProfileException(body['name'])
+                raise
+
+            self._associate_all_services(s, profile.tenant_uuid, profile_uuid, body['services'])
 
     def get(self, visible_tenants, profile_uuid):
-        filter_ = and_(
-            Profile.uuid == profile_uuid,
-            Profile.tenant_uuid.in_(visible_tenants),
-        )
+        filter_ = self._build_filter(visible_tenants, profile_uuid)
         with self.new_session() as s:
             profile = s.query(Profile).filter(filter_).first()
+            if not profile:
+                raise exception.NoSuchProfileAPIException(profile_uuid)
+
             return self._profile_schema.dump(profile).data
 
     def list_(self, visible_tenants, **list_params):
         filter_ = self._list_filter(visible_tenants, **list_params)
         with self.new_session() as s:
             query = s.query(Profile).filter(filter_)
-            # add pagination here
+            query = self._paginate(query, **list_params)
             return self._profile_schema.dump(query.all(), many=True).data
 
-    def _list_filter(self, visible_tenants, uuid=None, name=None, **list_params):
+    def _build_filter(self, visible_tenants, profile_uuid):
+        filter_ = Profile.uuid == profile_uuid
+        if visible_tenants is not None:
+            if not visible_tenants:
+                raise exception.NoSuchProfileAPIException(profile_uuid)
+            filter_ = and_(filter_, Profile.tenant_uuid.in_(visible_tenants))
+        return filter_
+
+    def _associate_all_services(self, session, tenant_uuid, profile_uuid, services):
+        for service_name, config in services.items():
+            service_uuid = self._create_service(session, service_name)
+            profile_service_uuid = self._create_profile_service(
+                session, tenant_uuid, profile_uuid, service_uuid, config,
+            )
+            for source in config['sources']:
+                source_uuid = str(source.get('uuid'))
+                self._create_profile_service_source(
+                    session, tenant_uuid, profile_service_uuid, source_uuid,
+                )
+
+    def _dissociate_all_services(self, session, profile_uuid):
+        session.query(ProfileService).filter(
+            ProfileService.profile_uuid == profile_uuid,
+        ).delete(synchronize_session=False)
+
+    def _list_filter(self, visible_tenants, uuid=None, name=None, search=None, **list_params):
         filter_ = text('true')
         if visible_tenants is not None:
             if not visible_tenants:
@@ -81,7 +124,29 @@ class ProfileCRUD(BaseDAO):
             filter_ = and_(filter_, Profile.uuid == uuid)
         if name is not None:
             filter_ = and_(filter_, Profile.name == name)
+        if search is not None:
+            pattern = '%{}%'.format(search)
+            filter_ = and_(filter_, Profile.name.ilike(pattern))
         return filter_
+
+    @staticmethod
+    def _paginate(query, limit=None, offset=None, order=None, direction=None, **ignored):
+        if order and direction:
+            field = None
+            if order == 'name':
+                field = Profile.name
+
+            if field:
+                order_clause = field.asc() if direction == 'asc' else field.desc()
+                query = query.order_by(order_clause)
+
+        if limit is not None:
+            query = query.limit(limit)
+
+        if offset is not None:
+            query = query.offset(offset)
+
+        return query
 
     @staticmethod
     def _create_profile(session, body):
@@ -91,22 +156,27 @@ class ProfileCRUD(BaseDAO):
             tenant_uuid=body['tenant_uuid'],
             name=body['name'],
             display_uuid=display_uuid,
+            display_tenant_uuid=body['tenant_uuid'],
         )
         session.add(profile)
         try:
             session.flush()
         except exc.IntegrityError as e:
-            if extract_constraint_name(e) == 'dird_profile_display_uuid_fkey':
+            constraint_name = extract_constraint_name(e)
+            if constraint_name == 'dird_profile_display_uuid_tenant_fkey':
                 raise exception.NoSuchDisplay(display_uuid)
+            elif constraint_name == 'dird_profile_tenant_name':
+                raise exception.DuplicatedProfileException(body['name'])
             raise
         return profile.uuid
 
     @staticmethod
-    def _create_profile_service(session, profile_uuid, service_uuid, config):
+    def _create_profile_service(session, tenant_uuid, profile_uuid, service_uuid, config):
         config = dict(config)
         config.pop('sources', None)
         profile_service = ProfileService(
             profile_uuid=profile_uuid,
+            profile_tenant_uuid=tenant_uuid,
             service_uuid=service_uuid,
             config=config,
         )
@@ -115,19 +185,21 @@ class ProfileCRUD(BaseDAO):
         return profile_service.uuid
 
     @staticmethod
-    def _create_profile_service_source(session, profile_service_uuid, source_uuid):
+    def _create_profile_service_source(session, tenant_uuid, profile_service_uuid, source_uuid):
         if source_uuid is None:
             raise exception.NoSuchSource(source_uuid)
 
         profile_service_source = ProfileServiceSource(
             profile_service_uuid=profile_service_uuid,
+            profile_tenant_uuid=tenant_uuid,
             source_uuid=source_uuid,
+            source_tenant_uuid=tenant_uuid,
         )
         session.add(profile_service_source)
         try:
             session.flush()
         except exc.IntegrityError as e:
-            if extract_constraint_name(e) == 'dird_profile_service_source_source_uuid_fkey':
+            if extract_constraint_name(e) == 'dird_profile_service_source_source_uuid_tenant_fkey':
                 raise exception.NoSuchSource(source_uuid)
             raise