+++ wazo_dird/plugins/api/http.py
@@ -1,5 +1,4 @@
-
-# Copyright 2016-2018 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import logging
@@ -25,13 +24,18 @@ class ApiResource(ErrorCatchingResource):
             try:
                 plugin_package = module.module_name.rsplit('.', 1)[0]
                 spec = yaml.load(resource_string(plugin_package, self.api_filename))
-                specs.append(spec)
+                if not spec:
+                    logger.debug('plugin has no API spec: %s', plugin_package)
+                else:
+                    specs.append(spec)
             except ImportError:
                 logger.debug('failed to import %s', plugin_package)
             except IOError:
                 logger.debug('API spec for module "%s" does not exist', module.module_name)
             except IndexError:
                 logger.debug('Could not find API spec from module "%s"', module.module_name)
+            except NotImplementedError:
+                logger.debug('Are you sure you have an __init__ file in your module "%s"?', module.module_name)
         api_spec = ChainMap(*specs)
 
         if not api_spec.get('info'):